Added back dimmer and dpms splitted implementations. Added support for new Clightd camera settings "Sensor.Capture" parameter.
