+++ Extra/clight.conf
@@ -38,6 +38,12 @@
 ## Sensor device to be used (Webcam or ALS device)
 # sensor_devname = "video0";
 
+## Sensor settings to be used. You should set values taken from <linux/v4l2-controls.h>,
+## as "A=b,C=d" string, where A and C are v4l2 operations, and b and d are their requested values.
+## Example string below does read as "V4L2_CID_EXPOSURE_AUTO=V4L2_EXPOSURE_MANUAL,V4L2_CID_EXPOSURE_ABSOLUTE=166".
+## This is only useful for webcam sensors right now. Leave empty/commented to use webcam default values.
+# sensor_settings = "10094849=1,10094850=166";
+
 ## Screen syspath to be use
 # screen_sysname = "intel_backlight";
