+++ TODO.md
@@ -165,7 +165,8 @@ Eg: on new location we should send a message with new lat and lon and LOCATION m
 - [x] Update API wiki for screen compensation (#84)
 - [x] Update API wiki for inhibit_autocalib (#81)
 - [x] Update API wiki dropping NoInhibit (#76)
-- [ ] Update API wiki with new Load/Unload functions
+- [x] Update API wiki with new Load/Unload functions
+- [ ] Update API wiki with new SensorSettings writable property
 
 ### 4.X
